refactor(sentry): extract the autofix selector's GitHub I/O layer
Pure move ahead of the family-dedupe work (#1810). The selector sat at ~800
lines, above the 600-line soft cap, and the dedupe change adds ~100 more, so
the checklist's split-not-append rule applies. Lift every `gh` read
(defaultRunGh, listCodeFixStubs, listHandledShortIds, readStub,
openAutofixPrExists with its fork-PR fence, isOwnHeadPr, HEAD_OWNERSHIP_FIELDS)
plus the queue-scoping constants (LOCAL_SENTRY_PROJECT, AUTOFIX_SELECT_LABEL,
parseProject) into scripts/sentry-autofix-queue-io.mjs. The selector keeps the
pure parse / evaluate / collapse-orchestration / report / CLI layers and imports
the I/O directly; tests re-point their imports, no re-export shim. No behaviour
change.
